prints id when connects to ipfs
